+++ Lib/Composition/Composition.cs
@@ -478,7 +478,7 @@ namespace Lib.Composition
         {
             InitDiskCache();
             _forbiddenDependencyUpdate = bCommand.NoUpdate.Value;
-            _mainBuildResult = new MainBuildResult(!bCommand.Fast.Value, bCommand.VersionDir.Value);
+            _mainBuildResult = new MainBuildResult(!bCommand.Fast.Value, bCommand.VersionDir.Value, bCommand.SpriteVersionDir.Value);
             var proj = SetMainProject(PathUtils.Normalize(Environment.CurrentDirectory));
             proj.SpriteGeneration = bCommand.Sprite.Value;
             var start = DateTime.UtcNow;
@@ -606,7 +606,7 @@ namespace Lib.Composition
         void RunTest(TestCommand testCommand)
         {
             InitDiskCache();
-            _mainBuildResult = new MainBuildResult(false, testCommand.VersionDir.Value);
+            _mainBuildResult = new MainBuildResult(false, testCommand.VersionDir.Value, testCommand.SpriteVersionDir.Value);
             InitTestServer();
             InitMainServer();
             var proj = SetMainProject(PathUtils.Normalize(Environment.CurrentDirectory));
@@ -791,7 +791,7 @@ namespace Lib.Composition
         {
             IfEnabledStartVerbosive();
             InitDiskCache();
-            _mainBuildResult = new MainBuildResult(false, findUnusedCommand.VersionDir.Value);
+            _mainBuildResult = new MainBuildResult(false, findUnusedCommand.VersionDir.Value, findUnusedCommand.SpriteVersionDir.Value);
             var proj = SetMainProject(PathUtils.Normalize(Environment.CurrentDirectory));
             _logger.WriteLine("Build started " + proj.Owner.Owner.FullPath, ConsoleColor.Cyan);
             try
@@ -875,7 +875,7 @@ namespace Lib.Composition
             }
 
             InitDiskCache(true);
-            _mainBuildResult = new MainBuildResult(false, command.VersionDir.Value);
+            _mainBuildResult = new MainBuildResult(false, command.VersionDir.Value, command.SpriteVersionDir.Value);
             InitTestServer();
             InitMainServer();
             SetMainProject(PathUtils.Normalize(Environment.CurrentDirectory)).SpriteGeneration = command.Sprite.Value;